agent/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentsuite-agent"
version = "0.1.0"
description = "LangGraph backend workspace for AgentSuite."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12,<1.0.0",
  "langchain>=1.2.15,<2.0.0",
  "langgraph>=1.1.9,<2.0.0",
  "openai>=1.76.2,<2.0.0",
  "python-dotenv>=1.2.2,<2.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "uvicorn>=0.34.2,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5,<9.0.0",
]

[project.scripts]
agentsuite-agent = "agentsuite_agent.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]
